added geothermal district cooling
